Bug fixes
Overhaul Util functions and added experimental feature
