+++ main.py
@@ -1,4 +1,5 @@
-from Util_PMP import cardCreated, cardMovedUp, cardOnHold, cardErraticMove, cardSkippedUp, calculatePoint, counter, resetCounter,config, performanceChart
+from Util_Embed import performanceChart, cardCreated, cardMovedUp, cardOnHold, cardSkippedUp, cardErraticMove
+from Util_PMP import calculatePoint, counter, config
 from datetime import datetime, timedelta
 from trello import TrelloClient
 from discord.ext import tasks
@@ -18,7 +19,7 @@ async def checkChange(aDict, card, channel):
         aDict (DICT): Dictionary containing card information
         card (OBJ): Card object
     """
-    author =  config.clientTrello.get_member(card.idMembers[0]).username if card.idMembers else "Unknown"
+    author =  config.Trello.get_member(card.idMembers[0]).username if card.idMembers else "Unknown"
 
     if  aDict.get("destination").get("name") == "On Hold":
         await cardOnHold(card.name.upper(), channel, author)
@@ -42,16 +43,14 @@ async def checkCardMovement(listID, targetTime, parentName):
     """
     for card in listID:
         last_move = card.list_movements() 
-        print(card.name)
         counter(parentName)
         try:
             if (last_move[0].get('datetime').replace(tzinfo=pytz.UTC) > targetTime):
-                print("Change Detected")
                 await checkChange(last_move[0], card, bot.get_channel(config.channelID))
 
         except IndexError:
             None
-        await asyncio.sleep(0.05)
+        await asyncio.sleep(0.09)
         
         
 async def checkCardCreation(listID, targetTime):
@@ -62,12 +61,10 @@ async def checkCardCreation(listID, targetTime):
         listID (INT): Trello Board ID
     """
     for card in listID:
-        print(card.name)
         if (card.created_date > targetTime):
-            print("Change Detected")
             await cardCreated(card.name, bot.get_channel(config.channelID))
             
-        await asyncio.sleep(0.05)
+        await asyncio.sleep(0.09)
         
 
 async def checkTrello(targetTime):
@@ -77,12 +74,10 @@ async def checkTrello(targetTime):
     Args:
         targetTime (datetime): Main time comparison used for calculation
     """
-    resetCounter()
-    print("#########Checking Boards#########")
+    config.resetCounter()
     for boardID, parentName in config.board_MVLS.items():
         await checkCardMovement((config.Trello.list_boards()[-1]).get_list(boardID).list_cards(), targetTime, parentName)
 
-    print("#########Checking Creations#########")
     for boardID in config.board_CTLS:
         await checkCardCreation((config.Trello.list_boards()[-1]).get_list(boardID).list_cards(), targetTime)
 
@@ -92,16 +87,28 @@ async def checkTime():
     """
     Main Loop function used for capturing and saving time values used for comparisons
     """
-    print("checkTime")
     if (datetime.now(pytz.timezone("Etc/GMT+0")) > config.targetTime):
         await  performanceChart(bot.get_channel(config.channelID))
         config.targetTime += timedelta(hours = 24)
+        
+    elif (config.countW != 0 and config.reportExecute is True):
+        await  performanceChart(bot.get_channel(config.channelID))
+        config.reportExecute = False
+        
     else:
         start_time = time.time()
         await checkTrello((datetime.now(pytz.timezone(config.timezone)) - timedelta(0,config.executionTime)).replace(tzinfo=pytz.UTC))
         result = (time.time() - start_time)
         config.executionTime = result
         print(f"#########{result}#########")
+        
+        
+@bot.event
+async def on_message(message):
+    if message.content.startswith('!Report'):
+        config.reportExecute = True
+        await bot.get_channel(config.channelID).send("Request sent. The request could take a while, please be patient.")
+                 
 
 @bot.event
 async def on_ready():
@@ -110,7 +117,6 @@ async def on_ready():
     """
     print("I`m Alive")
     await bot.change_presence(status=discord.Status.online, activity=discord.Game(name="with your feelings", description =''))
-    print(config.targetTime, datetime.now(pytz.timezone("Etc/GMT+0")))
     checkTime.start()
